FAC-609 Move 'InvalidContentException' to common module
