+++ flickit-assessment-common/src/main/java/org/flickit/assessment/common/exception/api/ErrorCodes.java
@@ -4,6 +4,7 @@ public class ErrorCodes {
 
     public static final String NOT_FOUND = "NOT_FOUND";
     public static final String INVALID_INPUT = "INVALID_INPUT";
+    public static final String INVALID_CONTENT = "INVALID_CONTENT";
     public static final String INTERNAL_ERROR = "INTERNAL_ERROR";
     public static final String ACCESS_DENIED = "ACCESS_DENIED";
     public static final String UNAUTHORIZED = "UNAUTHORIZED";
