+++ flickit-assessment-common/src/main/java/org/flickit/assessment/common/exception/handler/InvalidContentExceptionHandler.java
@@ -1,9 +1,9 @@
-package org.flickit.assessment.kit.adapter.in.rest.exception;
+package org.flickit.assessment.common.exception.handler;
 
 import lombok.extern.slf4j.Slf4j;
 import org.flickit.assessment.common.application.MessageBundle;
+import org.flickit.assessment.common.exception.InvalidContentException;
 import org.flickit.assessment.common.exception.api.ErrorResponseDto;
-import org.flickit.assessment.kit.application.exception.InvalidContentException;
 import org.springframework.core.Ordered;
 import org.springframework.core.annotation.Order;
 import org.springframework.http.HttpStatus;
@@ -12,7 +12,7 @@ import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 
-import static org.flickit.assessment.kit.adapter.in.rest.exception.api.ErrorCodes.INVALID_DSL_CONTENT;
+import static org.flickit.assessment.common.exception.api.ErrorCodes.INVALID_CONTENT;
 
 @Slf4j
 @RestControllerAdvice
@@ -24,6 +24,6 @@ public class InvalidContentExceptionHandler {
     @ResponseStatus(HttpStatus.BAD_REQUEST)
     ErrorResponseDto handle(InvalidContentException ex) {
         log.error("Invalid content error", ex);
-        return new ErrorResponseDto(INVALID_DSL_CONTENT, MessageBundle.message(ex.getMessage()));
+        return new ErrorResponseDto(INVALID_CONTENT, MessageBundle.message(ex.getMessage()));
     }
 }
