+++ flickit-assessment-kit/src/main/java/org/flickit/assessment/kit/adapter/in/rest/exception/api/ErrorCodes.java
@@ -6,7 +6,6 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
 public class ErrorCodes {
 
-    public static final String INVALID_DSL_CONTENT = "INVALID_DSL_CONTENT";
     public static final String UNSUPPORTED_DSL_CONTENT_CHANGE = "UNSUPPORTED_DSL_CONTENT_CHANGE";
     public static final String DSL_SYNTAX_ERROR = "DSL_SYNTAX_ERROR";
 }
