+++ flickit-assessment-kit/src/main/java/org/flickit/assessment/kit/application/service/DslTranslator.java
@@ -4,8 +4,8 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
+import org.flickit.assessment.common.exception.InvalidContentException;
 import org.flickit.assessment.kit.application.domain.dsl.AssessmentKitDslModel;
-import org.flickit.assessment.kit.application.exception.InvalidContentException;
 
 import static org.flickit.assessment.kit.common.ErrorMessageKey.TRANSLATE_KIT_DSL_UNABLE_TO_PARSE_JSON;
